[PAL/Linux-SGX] Mitigate CVE-2022-21166 aka INTEL-SA-00615
A malicious OS could map MMIO memory into the untrusted memory space of
an enclave application (which can be used for e.g. parameter passing
to/from ECALLs and OCALLs, or any external buffers that an enclave might
use to communicate with its application). Then when the enclave writes
to this memory, it could unintentionally propagate the stale (secret)
data in its core fill buffers into the uncore, where it could later be
extracted by malicious software.

Naturally aligned 8-byte writes are not affected by this vulnerability,
so as SW mitigation, this commit makes all untrusted memory writes
8-byte wide and naturally aligned.

There is a side effect introduced to the Exitless (aka RPC-threads aka
Switchless) feature by the fix of CVE-2022-21233 aka INTEL-SA-00657,
which marks the feature as insecure and didn't fix it due to its complex
usage of the untrusted memory. We assume the same for this fix.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>