[PAL/Linux-SGX] Retry when SGX driver's PF handler fails to add a page
Previously, Gramine expected no memory fault would happen inside
Linux-SGX PAL. But this was not always true during EDMM's page
allocation. In this flow, when PAL runs `ENCLU[EACCEPT]` on a new
enclave page which triggers a page fault (#PF), the SGX driver will
catch this #PF and try to handle it by issuing `ENCLS[EAUG]` for the
page. This handling of dynamically adding the enclave page may however
fail unexpectedly which will in turn lead to a memfault inside PAL.

This commit adds a retry logic for such case in Linux-SGX PAL's
exception handler.

Signed-off-by: Kailun Qin <kailun.qin@intel.com>